feat: Fix mypy issues
